+++ python/cell_object.cpp
@@ -196,6 +196,21 @@ static PyObject* cell_object_bounding_box(CellObject* self, PyObject* args) {
     return Py_BuildValue("((dd)(dd))", min.x, min.y, max.x, max.y);
 }
 
+static PyObject* cell_object_convex_hull(CellObject* self, PyObject* args) {
+    Array<Vec2> points = {0};
+    self->cell->convex_hull(points);
+    npy_intp dims[] = {(npy_intp)points.count, 2};
+    PyObject* result = PyArray_SimpleNew(2, dims, NPY_DOUBLE);
+    if (!result) {
+        PyErr_SetString(PyExc_MemoryError, "Unable to create return array.");
+        return NULL;
+    }
+    double* data = (double*)PyArray_DATA((PyArrayObject*)result);
+    memcpy(data, points.items, sizeof(double) * points.count * 2);
+    points.clear();
+    return (PyObject*)result;
+}
+
 static PyObject* cell_object_flatten(CellObject* self, PyObject* args, PyObject* kwds) {
     int apply_repetitions = 1;
     const char* keywords[] = {"apply_repetitions", NULL};
@@ -712,6 +727,7 @@ static PyMethodDef cell_object_methods[] = {
     {"area", (PyCFunction)cell_object_area, METH_VARARGS, cell_object_area_doc},
     {"bounding_box", (PyCFunction)cell_object_bounding_box, METH_NOARGS,
      cell_object_bounding_box_doc},
+    {"convex_hull", (PyCFunction)cell_object_convex_hull, METH_NOARGS, cell_object_convex_hull_doc},
     {"flatten", (PyCFunction)cell_object_flatten, METH_VARARGS | METH_KEYWORDS,
      cell_object_flatten_doc},
     {"copy", (PyCFunction)cell_object_copy, METH_VARARGS | METH_KEYWORDS, cell_object_copy_doc},