+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
 ## Unreleased
 ### Added
+- `Cell.convex_hull`
 - `Cell.filter` to remove elements by layer and data/text type
 - `FlexPath.path_spines()` and `RobustPath.path_spines()`
+### Changed
+- Bounding box calculations use the convex hull for efficiency
 
 ## 0.3.2 - 2021-02-15
 ### Fixed